Move DialogBox and MainWindow into gui folder
