+++ src/main/java/ari/gui/DialogBox.java
@@ -1,4 +1,4 @@
-package ari;
+package ari.gui;
 
 import java.io.IOException;
 import java.util.Collections;
